почему это не работает? может ли быть такое что у меня что-то не поломалось в самом окружении? flask_sqlalchemy например как то слетел?
