+++ webapp/models.py
@@ -1,0 +1,64 @@
+from datetime import datetime
+from flask_login import UserMixin
+from flask_sqlalchemy import SQLAlchemy
+from werkzeug.security import check_password_hash, generate_password_hash
+
+db = SQLAlchemy()
+
+
+class News(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String, nullable=False)
+    url = db.Column(db.String, unique=True, nullable=False)
+    published = db.Column(db.DateTime, nullable=False)
+    text = db.Column(db.Text, nullable=True)
+
+    def __repr__(self):
+        return '<News {} {}>'.format(self.title, self.url)
+
+
+class User(db.Model, UserMixin):
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(50), unique=True)
+    password = db.Column(db.String(500), nullable=True)  # не пустое, под хеш пароля
+    role = db.Column(db.String(10), index=True)
+    firstname = db.Column(db.String(50))
+    lastname = db.Column(db.String(50))
+    city = db.Column(db.String(100))
+    date = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def set_password(self, password):
+        self.password = generate_password_hash(password)
+
+    def check_password(self, password):
+        return check_password_hash(self.password, password)
+
+    @property
+    def is_admin(self):
+        return self.role == 'admin' or self.role == 'root'
+
+    def __str__(self):
+        return f'<users {self.id, self.email, self.firstname, self.lastname, self.password}>'
+
+    def __repr__(self):
+        return {'id': self.id, 'email': self.email, 'password' : self.password, 'role': self.role}
+
+
+class GPU(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    citilink_id = db.Column(db.Integer, nullable=False)
+    category_id = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.Float, nullable=False)
+    old_price = db.Column(db.Float, nullable=False)
+    short_name = db.Column(db.String, nullable=False)
+    category_name = db.Column(db.String, nullable=False)
+    brand_name = db.Column(db.String, nullable=False)
+    club_price = db.Column(db.String, nullable=True)
+    picture = db.Column(db.String, nullable=False)
+
+    def __repr__(self):
+        return f'<GPU: id:{self.id}\n, shortname: {self.shortName}\n, price: {self.price}>'
+
+
+if __name__ == "__main__":
+    pass
